content-security-policy/frame-src/frame-src-same-document-meta.sub.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<html>
  <body></body>
  <script>
    async_test((test) => {
      // 1. Load an iframe (not blocked).
      let iframe = document.createElement("iframe");
      iframe.name = "theiframe";
      iframe.src =
        "http://www1.{{host}}:{{ports[http][0]}}/content-security-policy/frame-src/support/frame.html?0";
      iframe.onload = test.step_func(() => {
        // 2. Start blocking iframes using CSP frame-src 'none'.
        let meta = document.createElement("meta");
        meta.httpEquiv = "Content-Security-Policy";
        meta.content = "frame-src 'none'";
        document.getElementsByTagName("head")[0].appendChild(meta);

        // 3. Blocked same-document navigation using iframe.src.
        window.addEventListener(
          "securitypolicyviolation",
          test.step_func(() => {
            assert_true(
              true,
              "Failed to block same-document navigation in an iframe blocked by CSP frame-src dynamically using the <meta> tag when using iframe.src."
            );
          }),
          { once: true }
        );

        iframe.src =
          "http://www1.{{host}}:{{ports[http][0]}}/content-security-policy/frame-src/support/frame.html?1";

        // 4. Blocked same-document navigation using window.open.
        window.addEventListener(
          "securitypolicyviolation",
          test.step_func(() => {
            assert_true(
              true,
              "Failed to block same-document navigation in an iframe blocked by CSP frame-src dynamically using the <meta> tag when using window.open."
            );
          }),
          { once: true }
        );
        window.open(
          "http://www1.{{host}}:{{ports[http][0]}}/content-security-policy/frame-src/support/frame.html?2",
          "theiframe"
        );

        // 5. Regression test for https://crbug.com/1018385. The browser should
        // not crash while displaying the error page.
        test.step_timeout(test.done, 1000);
      });
      document.body.appendChild(iframe);
    });
  </script>
</html>
