LayoutTests/imported/w3c/web-platform-tests/content-security-policy/frame-src/frame-src-same-document-meta.sub-expected.txt

Blocked access to external URL http://www1.localhost:8800/content-security-policy/frame-src/support/frame.html?0
CONSOLE MESSAGE: Refused to load http://www1.localhost:8800/content-security-policy/frame-src/support/frame.html?1 because it does not appear in the frame-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

TIMEOUT Same-document navigations in an iframe blocked by CSP frame-src dynamically using the <meta> tag Test timed out
